--- include/CompraProd (2).h
@@ -1,24 +1,0 @@
-#include <iostream>
-#include <set>
-#include <string>
-#include <vector>
-#include <memory>
-#include <sstream>
-#include "Producto.h"
-
-class CompraProd{
-private:
-    int cantidad;
-    bool enviado;
-    Producto *productos; //REVISAR
-public:
-    //constructor
-    CompraProd();
-
-    //metodos de acceso
-    int getCantidad();
-    bool getEnviado();
-
-    //destructor
-    virtual ~CompraProd(){}; 
-};
