+++ include/TCategoria.h
@@ -1,0 +1,7 @@
+#include <iostream>
+
+enum TCategoria {
+    ropa,
+    electrodomesticos,
+    otros
+};
